+++ sageortus/src/ortus/resources/createmovies.sql
@@ -9,7 +9,7 @@ create table sage.movies (
     alternate_name varchar,
     url varchar,
     votes int,
-    rating float,
+    rating int,
     tagline varchar,
     certification varchar,
     releasedate date,
